Improving systemctl integration
When the AO restarts, the script was throwing a disconnect event but not exiting. So the soda machine pi had to be restarted every time the AO was restarted. To fix this, removed the thrown error, but then it was having multiple "Connected with authentication!!!" lines. This was because the authentication handler was being re-initialized every time it reconnected. These handlers don't need to be nested. Now it reconnects automatically without exiting.